Update patch set 1
Patch Set 1:

Lots of info here that I didn't know about. Thank you.

> > 
> > When the update rect comes directly from the app server, drawing is already clipped to that [...] so constraining the clipping area to that update rect does nothing.
> 
> No, it isn't. app_server will clip the drawing to the visible area of your BView (so you don't draw outside the view). But it will tell your application which part absolutely needs to be redrawn (by passing you an UpdateRect).
> 

I've written a simple BView that just logs the calls to Draw, its updateRect and the clipping region. When I call Invalidate on a small rectangle, the clipping region is set to that rectangle, even though the full view is exposed. When I hide part of the view putting some windows over it and then expose part of that closing one of the windows, the clipping region contains just the newly exposed parts, even though the visible area was larger. In both cases updateRect is just the rect containing all the rects in the clipping region. This is in line with my understanding of this from The Be Book: "When drawing to the screen, the updateRect is the smallest rectangle that encloses the current clipping region for the view."

Now, I admit I've not played a lot with it, tested different flags, views hierarchies or other stuff that may change results.


> [...]
> So, there is a common use case here where you need to make sure no drawing happens outside of a given rectangle, given in non-transformed space.
> 
> If you use BDirectWindow, there is no clipping at all done by app_server, and you must be careful to do your own. However, app_server provides you with a BRegion where you are allowed to draw.
> 
> For these two cases, I think it makes sense to use GetClippingRegion/ConstrainClippingRegion.
> 

I agree and don't dispute that. By all means use whatever function makes most sense for the circumstances you are in.

> > 
> > When the view is doing its stuff and wants to further restrict where the control look can draw, I say it's the caller's responsibility to clip. [...] If I want a circle but only the part within some shape [...]
> 
> If you want to do this in an application, you need to use ClipToShape or ClipToPicture (and their inverse versions). These two should follow view transforms.
> 

Sure. And they do. But again, it's not about what clipping function to call. Let me state my problem in a possibly clearer way: the question I'm into now is whether the update rect passed to the ControlLook methods means the caller is saying (a) "for whatever reason you must only draw inside this rect, I've not clipped to that and it's now your responsibility" or (b) "for whatever reason what you draw outside this rect won't go anywhere, so you may be more efficient if you do some checks and avoid drawing some stuff", and what coordinate system it is in.

Option (a) makes no sense to me. My stance is that if you are calling a function to draw something (be that within or outside Draw, from a different thread, in a BDirectWindow that I don't know how it works, or wherever) and you need that to be restricted, it is you, the caller, who must do the clipping, and it doesn't matter if the function is FillEllipse, DrawButtonFrame or DrawFairyWithCastleInTheBackground. If the drawing function makes something complex it will probably do its own clipping, but for its own sake in what it is drawing, not as a result of its caller telling it must clip this or that.

If it is (a) we can stop here. Nothing of what I thought and did makes any sense.

If it is (b) clipping to updateRect in ControlLook does nothing and we can go on talking about the coordinate system.


> > [...] It looks easier to me that the transformation, if necessary, is done by the caller
> 
> You cannot do this correctly, if the transform has any rotation or shearing [...] cannot be represented by just a top left and a bottom right corner.
> 

Right. Just as you cannot represent a general BRegion with just one BRect. Being an optimization hint, all you need is a rect that contains all the clipping region.

There's a point against using transformed coordinates for that rect, though. If you are not doing anything special and the update rect is the one provided to your View or you do have more clipping but it is done is screen coordinates, it's easier for the caller to directly use that and not have to transform anything, of course. So you may prefer that on the basis of a cleaner API for some cases. Personally, I find it unnatural that you pass the drawing rect in one system and the updateRect in a different one, and a waste to do the transformation in every call to ControlLook when there may be many for the same state of the view, which may have had to do yet another transformation if the update info it had was in draw coordinates.

Unlike the first part of the question, all I need to accept the option that I don't like in this one is that someone experienced say that's The Right Thing.


> Moreover, you can apply [and restore] transforms while you are inside BView::Draw. [...] So you can't even assume the existence of a single "transformed" coordinate space during the drawing operation at all. [...]

I'm lost here. Unless you understood I was asking to change what BView::Draw gets as updateRect, I can't make any sense of it. This is about the update rect passed to ControlLook methods, and there is a transformed coordinate space when you call any of them, or they wouldn't know where to draw.

I think the rest of that paragraph goes with the same assumption of me wanting to change what everything gets as updateRect if there's one in their signature or something in app_server inner workings. Or maybe they were rhetorical questions while explaining why those are like they are. Not my intention to change any of those. And being present in BeOS, I guess we can't. 

> 
> I don't know which side of the fence ClipToRect/ClipToInverseRect should fall on. We should just pick one and document it. The existence of an "inverse" function makes me think it should be like ClipToShape and ClipToPicture. And that also makes it consistent with FillRect/StrokeRect, jsut as ClipToShape is consistent with FillShape/StrokeShape.

I thought I had seen it documented that ConstrainClippingRegion is not affected by transforms and that calls are not cumulative, whereas ClipToRect/Picture/Shape do use transforms and their calls are cumulative, but I can't find it. Unless there are several paths, ClipTo[Inverse]Rect works as you say: https://git.haiku-os.org/haiku/tree/src/servers/app/DrawState.cpp#n474. In fact, if the transform won't give it a rect, it just builds a rectangular shape and calls ClipToShape.

Patch-set: 1
Reviewer: Gerrit User 1000223 <1000223@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>