Update patch set 2
Patch Set 2:

(1 comment)

The only other warnings blocking Werror for jpeg2000 are related to a compilation flag for jasper.
The cmake configuration of jasper does not check for the gcc compiler version. That particular flag is supported from 4.9 upwards only. I have no idea to fix that on our side. Creating a modified repo seems not worth it.

Patch-set: 2