Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The style is a bit behind, right.
> I'd like opinions on the colorspace ids. Should they be marked experimental also?

My understanding from looking at the constants is that 0x1000 is the mask for big endian, and 0x2000 is the mask for the alpha channel? And the low bits seem a bit arbitrary:

1 - 1bpp
2 - 8bpp (grey)
4 - 8bpp (palette)
10 - 15bpp
5 - 16bpp
3 - 24bpp
8 - 32bpp

And that's only for RGB, the other colorspaces don't really match. So, there doesn't really seem to be any logic to this. Let's just pick some values that are available?

As for making them experimental, if you have the colorspace conversion, I think this is the same level of support as the non-RGB colorspace (no way to draw on them using BView). It is enough to work with translators and media files and eventually convert them to RGB32 for display.

Patch-set: 1