Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Fixing it would make the diffs easier to read.  Probably will have to nuke the Gerrit changes after this one and recreate them.

That is not how you're supposed to use Gerrit.

Please include the change-id in the commit messages. It is listed in Gerrit below the commit message, for this one it's this line:

Change-Id: I9c660b1569b5439f2cd9088fde4e96512e7817d8

When you push a commit for review and the commit message includes this line, Gerrit will automatically notice that this is an update to the previous version, and update the change request. This way, people can easily compare different versions of a commit, and follow all the discussion.

No need to nuke anything 😊

If you have set up your Haiku source checkout following the instructions (https://www.haiku-os.org/guides/building/get-source-git), you should have done this:

    scp -p <login>@git.haiku-os.org:hooks/commit-msg "haiku/.git/hooks/"

This installs a git hook locally that makes sure all the commits you create are automatically assigned a change-id, so that you don't need to copy it back from the gerrit web interface.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>