Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> The correct way to do this (with gcc) is using unions. See https://exceptionshub.com/unions-and-type-punning.html for examples

Thank you for the excellent reference regarding this topic. From my point of view the relevant quotes from the provided source are:

„To re-iterate, type-punning through unions is perfectly fine in C (but not in C++).“

„C11 section 6.5.2.3 §3:
A postfix expression followed by the . operator and an identifier designates a member of a structure or union object. The value is that of the named member
with the following footnote 95:
If the member used to read the contents of a union object is not the same as the member last used to store a value in the object, the appropriate part of the object representation of the value is reinterpreted as an object representation in the new type as described in 6.2.6 (a process sometimes called ‘‘type punning’’). This might be a trap representation.“

„GNU extensions to standard C++ (and to C90) do explicitly allow type-punning with unions. Other compilers that don’t support GNU extensions may also support union type-punning, but it’s not part of the base language standard.“

Therefore, by using unions we leave the C++ standard and rely on compiler-specific extensions. I would have preferred to remain compliant with the standard.

Patch-set: 3