Update patch set 1
Patch Set 1: Code-Review-1

I'm not fond of doing it like this. The app_server is already built as a real server, with the ability to serve any number of users. If it's just a temporary solution, I wouldn't permanently remove code in AppServer.cpp, and add comments to explain the situation.
I think it also makes a lot of sense to keep app_server, and input_server as global system services.

As you already mentioned, this breaks down with multi-user, so it's also not a future proof solution either. I would prefer to solve this one for real.

The obvious solution in this case is to let the input_server add-on launch applications not directly, but by using the user's launch daemon instead. This one can share the environment already without further effort.

Patch-set: 1
Reviewer: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1