Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > If it doesn't work very well yet (I think there are some known bugs?) we could put it in the haiku_extras package (or create more packages for other things...)
> 
> I'm good with that.  I think if we're going to be going with our "classic" style until after R1, we could at least spice it up a bit. (a common reviewer opinion is "our classic style isn't for everyone" lol)

These people know nothing about UI design and don't see that our design is better for usability and discoverability (really, modern UIs where everything is flat and you have no idea what is a button and what is a label are quite bad, and an over-reaction to Apple going too far in the opposite direction with skeumorphism).

I'm not against changes, but I won't buy it if the only reason is "personally I think it looks nicer". So it's a good idea to have two different themes (or a few more) and see which ones people
use and maybe conduct some UX research (we can set this up at some opensource conferences when they start happening again)

Patch-set: 1