Update patch set 1
Patch Set 1:

> Remote desktop is a good reason for optional (!) server side drawing support, that's true. But in the world of compositors, you usually only draw (client-side filled) bitmaps in a single 3D context.

Right, that is the way it is usually done on X11 et al. But I think there is a better way to do it with server-side rendering: use only one 3D context for all applications, inside app_server (not one per-application, saving resources), and redraw frames within v-sync windows, almost totally eliminating the need for double-buffering on a per-application level.

Most OSes cannot do this due to having made architectural choices that preclude it (or, on X11, not using server side drawing at all). We have an opportunity to go this route that other OSes don't, and I think there are good reasons to consider it.

> I don't see how 32-bit change anything in this regard, at least not in that direction.

Because it is difficult but possible to have app_server use up all of its 2GB address space on 32-bit as it is; supporting multiple users in one process would make that extremely easy. Multiple app_servers for multiple users is basically the only tenable thing on 32-bit systems due to address space usage. (This presumes PAE but that is the case on nearly all 32-bit systems, I believe, x86 or otherwise.)

> Sure, that would be a solution. However, this is just app_server torn into two pieces (which makes sense IMO), and not moving it completely to the current user, which is what this discussion is about.

Well, until we implement that, putting app_server entirely in user context before that is broken off seems to make some sense.

> This is greatly exaggerated, and bugs like that can usually be fixed easily. You also don't want to share stuff between apps running under the same user -- like a malware and your web browser. This doesn't really change anything.

Indeed you don't, but applications on a single user or session already have various methods of communicating, and will be permitted to continue using them in a multi-user environment -- files, the clipboard, messages, etc. By contrast, applications on separate users generally will not be permitted such communication, and so app_server suddenly becomes a very attractive target if it has access to send messages to and from applications on other users, etc.

> [... agreeing on the missing parts ...]

:)

> No, what's there should be completely functional. Multi-screen support (for a single user) is not implemented yet at all.

X512 showed it working on the forums with test_app_server (though you can't move applications between the virtual "screens.")

Patch-set: 1