Update patch set 7
Patch Set 7: Code-Review-2

> No, GCC knows aligned_malloc(), so this part of the change is useless.

Except I think it ignores these semantics when we use -fno-builtin, as we do for the kernel and I believe also parts of libroot. So the earlier version of this change was thus more correct, and we may as well leave in the attribute.

Patch-set: 7
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2