Update patch set 2
Patch Set 2:

> Nice one! I would definitely be interesting to do some benchmarks between those two versions on real hardware. I guess you can easily switch back when you set HAIKU_KERNEL_ARCH to x86_gcc2 again?

Not so easily due to places where the KERNEL_ARCH is assumed to not have any suffixes. We probably should fix that in the build system, but that is a different problem.

Though I may have to do something about it, as it appears the x86_64 build failed due to something like that...

Patch-set: 2
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>