Update patch set 7
Patch Set 7:

so.. part of what irked me on this is drm is a *MASSIVE* project.   If we try to work in feature branches, it'll never get merged since it touches so much. (and I have a severe allergy to reworking patchsets over, and over, and over, and over. It's a waste of time.)

Just like in riscv, we should merge a base to work off of.  Then everyone can experiment in their own branches and work on feature functionality under a similar design api.

This adds the "bare minimum" base for any and all DRM designs in a way which is friendly to our existing driver / accelerant design.

These ioctl's "existing" don't break any existing code, nor does it introduce instabilities.

X512 looked at amdgpu, and these are the card centric drm calls...

#define DRM_AMDGPU_GEM_CREATE 0x00
#define DRM_AMDGPU_GEM_MMAP 0x01
#define DRM_AMDGPU_CTX 0x02
#define DRM_AMDGPU_BO_LIST 0x03
#define DRM_AMDGPU_CS 0x04
#define DRM_AMDGPU_INFO 0x05
#define DRM_AMDGPU_GEM_METADATA 0x06
#define DRM_AMDGPU_GEM_WAIT_IDLE 0x07
#define DRM_AMDGPU_GEM_VA 0x08
#define DRM_AMDGPU_WAIT_CS 0x09
#define DRM_AMDGPU_GEM_OP 0x10
#define DRM_AMDGPU_GEM_USERPTR 0x11
#define DRM_AMDGPU_WAIT_FENCES 0x12
#define DRM_AMDGPU_VM 0x13
#define DRM_AMDGPU_FENCE_TO_HANDLE 0x14
#define DRM_AMDGPU_SCHED 0x15

That's not really the massive glacier people make it out to be.

Patch-set: 7
Reviewer: Gerrit User 1000000 <1000000@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0