Update patch set 7
Patch Set 7:

> Patch Set 7: Code-Review-2
> 
> 1. I don't see how this would let a Vulkan driver recognize anything as there is no actual DRM code in here.
> 
> 2. I actually think we don't need feature branches ... because we can keep the DRM drivers fully out of tree and in a separate repository entirely (at least this was my plan back when I looked into this before.) This will avoid quadrupling the size of the Haiku repository.
> 
> 3. Some of those IOCTLs are clearly marked "GEM", i.e., the DRM memory management subsystem. This by itself is absolutely massive, and those handful of ioctls are way more complicated than they may appear on first glance...


To summarize, the overall argument here is the following:

1) We should keep the drm stuff completely out of tree (not even sure how this would work since we need kernel drivers to manage the cards).   drm should 100% be a compatibility layer and all mode setting should happen via drm drivers (completely abandoning the driver / accelerant design).  I guess BAccelerant / BScreen / etc would be some kind of compatibility API's which translate into drm calls "somehow".

2) We should add hooks to our existing drivers for drm ioctl's and work within our existing driver / accelerant api.


The issue with 1 is it's complexity. We'll need to completely redesign the app_server, throw away all of our existing drivers, and start over. 

The issue with #2 is we'll need to take the complex drm per-card memory management code and implement it within our kernel drivers.


Until we decide if #1 is what we want to do, or #2 is what we want to do... nothing will get done for 3d acceleration.

Patch-set: 7
Reviewer: Gerrit User 1000000 <1000000@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0