Update patch set 2
Patch Set 2:

Hm.  So where do we think the "proper" fix should be?

* app_server
* radeon_hd driver

The misalignment is due to the lack of write combining on the current riscv64 PCI controller found in the Unmatched?

Is there anyway to (easily) disable write-combining on x86 PCI?  Theoretically if we disable write combining on x86, and the issue appears on vesa + radeon_hd, it points to an exposed bug in app_server. If the issue only appears in radeon_hd, then it's a bug in radeon_hd.

Patch-set: 2
CC: Gerrit User 1000000 <1000000@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>