Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hm.  So where do we think the "proper" fix should be?
> 
> * app_server
> * radeon_hd driver

The options are:
- Make sure this function is only called on 64-bit aligned data
- Make this function safe to use on unaligned data (that is, replace it with a memcpy)

> 
> The misalignment is due to the lack of write combining on the current riscv64 PCI controller found in the Unmatched?

No. The misalignment is due to the code not taking care of aligning things. On x86 this is not a problem, the CPU handles misaligned accesses just fine. On other architectures, this isn't always the case.

Write combining is just an extra optimization, allowing several small writes to consecutive bytes to be translated to one single larger write. But it needs the CPU to generate the correct accesses in the first place. Maybe write combining can hide the problem, but it's still better to do the right thing in the first place.

I did not check where the function is used. If it can copy things starting from any pixel position in a framebuffer, there is no way to make sure the data is properly aligned (if your even pixels are aligned, your odd ones aren't).

Patch-set: 2