Update patch set 2
Patch Set 2:

> Breaking the data down in reducing chunk sizes doesn't appear to make anything "faster".

The goal of this function appears to be to use 64bit copies for most of the transfer, supposedly that would be faster. Then copy the last few remaining bytes with smaller transfers as needed.

Again, we should just use memcpy which should be implemented in the most efficient way already. I think it is the third time I ask why we don't just use memcpy here. I think it's reasonable to expet that gcc will be just as good, if not better, at optimizing memcpy, and if it isn't, we should replace our memcpy implementation, not write custom memory copy functions everywhere.

Patch-set: 2