Update patch set 1
Patch Set 1:

(4 comments)

Thanks for your contribution!

Do we really need to keep the old look as an option? It doesn't seem so useful to me, why not allow only the new one?

I have also suggested some improvements to the implementation, see the inline comments.

It may also be nice to provide a screenshot of the changes (I have seen one in your Twitter post, but probably not everyone reviewing this will have come accross it). It will help people understand what you're doing and make reviewing the code easier.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>