Interface Kit: add default dark appearence color constants
This is mainly required for webkit.
A website can explicitly for controls request to draw them
with either a light or a dark mode. It is not required that
this matches the system color scheme, that means a website
can declare specific controls to always render in a light theme
even if the OS is in a dark mode or vice versa

Change-Id: I8added6ada702eb09e95f5f2d6e34b748067c6bc