Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I'm guessing this fixes https://dev.haiku-os.org/ticket/12890

It is needed for the fix (https://review.haiku-os.org/c/haiku/+/4457) to work, but does not by itself fix any ticket that I know of.

It should fix glitches you may have stacking several masks when at least one of them has the inverse bit, like in the clipping tests in the parent commit. For a real world thing, I've noticed the spinning widget shown in the forums while loading a page is fully visible with this patch instead of being cropped to a smaller rectangle, but I've not yet taken the effort to check how that is drawn.

Patch-set: 2