Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Why not do it in the constructor?

Um, do what exactly? Calculate the duration? Because that would always do it, while right now it's only done if the attribute does not exist. Most of the time that calculation will never be triggered, but in case of read-only volumes, it is always triggered and causes jank.

Patch-set: 2