Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Why do we have this function in libgnu, anyway? As all the BSDs have it and it seems pretty useful generally, perhaps we should just move it into libroot (and replace our implementation with musl's, which is more optimized?)

We can move it to libbsd under _DEFAULT_SOURCE instead of libgnu under _GNU_SOURCE. We shouldn't put it in libroot to avoid conflicts if someone provides a function called memmem that behaves differently, at least until it is standardized.

There seem to be discussions in progress to include it in POSIX issue 8: https://www.opengroup.org/austin/docs/austin_1110.pdf
https://www.austingroupbugs.net/view.php?id=1061

I hope this is released soon, but until then I don't think we should add this to libroot.

Patch-set: 2