Update patch set 1
Patch Set 1:

Ideally there would be few to no extra context switches in the case where only one user is active, like there is now.

I don't know when I would get to fixing up the Shortcuts addon. I think it deserves a bit of an overhaul and more features, anyway, but that's rather low on the priority list.

I'll make the one change to just if 0 out the code in AppServer.cpp instead of removing it, in case we revert back to global app_servers, but if you are not opposed, then maybe indeed it will not be needed, before merging.

Patch-set: 1