Update patch set 2
Patch Set 2: Code-Review+2

I have the same questions as usual, why is all this code in efi/arch/ when it ends up being identical for all architectures? Can't we move some of this to efi/ and implement it once for all architectures?

That being said, if this gets the ARM port further, we can merge it and refactor that later.

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2