Update patch set 5
Patch Set 5:

As far as I can tell you use UEFI to set the memory map?

If so it is something we should make general and replace in x86_64, risc-v and use for arm64 as well.
It is ok to leave it in arm as a first version though. Nice!

Patch-set: 5
Reviewer: Gerrit User 1000021 <1000021@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>