Update patch set 4
Patch Set 4: Code-Review+1

This fixes #16412 which is a bug comparing the translated decorator name to the untranslated one. It only applies to the "Default" decorator name as the other ones aren't translated yet, but if they are someday this will also make those work as it uses the file name of the decorator on disk in the comparison which won't change even if the name gets translated in Appearance preferences pop up.

Patch-set: 4
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+1