sparc: WIP MMU code
Start implementing the UltraSparc / UltraSparc II MMU. Later versions
are different (increase in number of physical bits, and then later on a
complete redesign) so we will need different implementations of the MMU
depending on the CPU type.

Nothing really useful yet, just the structure to be filled in and basic
data definitions. Also add some documentation on how the MMU works.

Change-Id: I33e07432c715bcd87a6f7b881d1901ee63217455