Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > > I kind of expect the malloc to be aligned?
> > 
> > That's clearly not the case, we can see it if we add some debug output:
> > 
> > kernel_args_malloc(): 2688 bytes (49154 bytes left)
> > kernel_args_malloc: returning 0x59808ffe
> 
> Yes, I wonder why? Perhaps we should make it return aligned memory. Someone that knows more about this?

This is a very simplified allocator used just to handle a small memory area where we store all data passed from the bootloader to the kernel. Initially there was not a lot of stuff to pass, and on x86 alignment does not matter anyway, so it was not designed with alignment.

If it is neeeded, we could add it, but if the problem happens only in this one case on ARM, maybe it's not worth complexifying the allocator?

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>