Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> We get a bunch of "bad data" and "unhandled pheader type" errors during bootup.
> 
> This change mirrors the content of hrev47798:
> * ignore PT_ARM_UNWIND
> * allow up to 32K of space between sections of an ELF file

PT_ARM_UNWIND is related to exception handling. We don't support exceptions in the kernel, so ideally, if possible, we should make sure that the unwind tables for exceptions are not generated at all. I think this should be possible by adding -fno-exceptions to the kernel and add-ons compiler flags? If that's already done qnd the compiler generates these anyway, the second thing to do is make sure these sections are ignored by the linker script. And finally, if none of this can work, then yes, we can ignore it here. But it means we are putting things in kernel add-ons that are not used, and so we are just wasting disk space.

As for the gap between sections, that could be solved by making the alignment smaller in the linker script, I think? I see no reason to have a gap larger than 1 page (4K) on ARM? There as well it looks like this would be wasting resources (a few kilobytes of RAM usage) for no reason.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>