package command: Enable zstd compression by default.
Note that zstd-compressed packages cannot be installed unless the
running system can already read them, even as part of a system
update transaction. On x86_64, where zstd has been enabled
for quite a while, this is not a problem; but on x86_gcc2 where
it has only been enabled since dedd2be949bb0859181ebc51ca2091dc8ba6bda4,
a week ago, this will be an issue; thus, it will be a problem for
beta3 to beta4 updates. (EFI bootloaders must be updated too,
but that is another problem.)