Update patch set 3
Patch Set 3:

It was discussed, and axeld said in the last comments that he could "live with this crutch for a while (and that he'd be potentially in favor of per-user app_servers anyway). PulkoMandy already said that a change like this was OK for now on the ticket. That's 3 in favor "for now" at least, and having waited about a month and gotten no other opposition besides your comments, it seemed there was enough of a consensus to merge it.

Patch-set: 3