Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > This will avoid future confusion between the
> virtual width/height and the actual display timings, if we implement
> scrolling someday.
> 
> Isn't scrolling already working? At least for me is is working after Rudolfc changes.

I have not tried it, and I don't know how it is implemented currently and if it works on all generations of hardware.

Anyway, same result: if it already works, this will probably help avoiding breaking it again by using the wrong fields by accident when setting the video mode.

Patch-set: 1