Update patch set 1
Patch Set 1: Code-Review-1

On ARM the interrupts are indeed described by 3 integers and not just one. This can be detected from the FDT by looking at the #interrupt-cells property. Ideally, we will need to get all 3 values from the FDT to properly configure the interrupt.

See here for a description of how interrupts are described in FDT: https://elinux.org/Device_Tree_Usage#How_Interrupts_Work

There is in fact a TODO about this down in the same function, it would be nice to solve it instead of adding hardcoded platform specific code here.

This code is actually misusing the "ord" parameter which, if I understand correctly, is supposed to mean "get me the infos for the Nth interrupt for this device", and not "get me the second integer in the interrupt property".

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1