sparc: WIP traps/interrupts handlers.
This will be needed to intercept MMU cache misses so we need to get it
up and running quite early.

For now, attempt to just forward all calls to the existing handlers set
up by openboot. However, this does not work. I'm not sure why, but maybe
it's because the new handlers are not locked into the iMMU cache and as
a result, MMU misses cannot be handled.

Change-Id: I5dc0950d336ed7ac837d1300882faac2d3f39d2c