Update patch set 1
Patch Set 1:

> I mean created date on Tracker info of already present file.

Right, that isn't implemented here. Guess there is a way to patch that.

> For some *.jpg photo files it shows "Is a directory" error.

Now that is strange. As you can see in the code, we return that if the directory flag is set on the inode. Maybe that check is actually wrong somehow...

Patch-set: 1
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>