Update patch set 1
Patch Set 1: Code-Review-1

This is not how we should implement this feature. There should be only 1 set of color constants not a duplicate set of light and dark and we should work the rest out in Appearance. It's nice to have a good set of dark color constants to use but we probably don't want to extend the API in this way.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1