Update patch set 1
Patch Set 1:

The major problem I noticed in testing is that for whatever reason the node monitoring just does not work. That is, e.g., if you rename a file in Tracker, it gets renamed in NTFS, but in Tracker it just totally disappears until you close and reopen the window.

I checked and verified that the node monitoring messages are being sent (e.g. adding message.Dump(dprintf); inside kernel/fs/node_monitor.cpp NotifyEntryMoved for instance) and they look correct to me. The code in rename at least seems to be functionally identical to what BFS does; I used it as a model when writing this code.

So, I am kind of baffled as to what the problem is there at present. Anyone have any ideas?

Otherwise, this just needs a lot more testing. I only verified the absolute basics (create, write) seem to work.

Patch-set: 1