Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Hi, just looked over this patch because I was asked to. 
> 
> So it's an update to the LVDS port code, afaict to add support for Ibex Peak which is H55/H57 chipset (first ever with PCH I read), where I did not see any LVDS port in the chipset from a quick search on internet (only HDMI, Display Port, DVI, VGA).
> Later generation 7 graphics, starting with the DDI interface, nolonger have LVDS ports either afaik.
> 
> Besides I think it's outdated seeing the changes  that were done to the driver in the meantime.
> 
> I'd say we forget about this patch therefore. That is, at least I myself don't have a H55/H57 system at my disposal, so if someone has one with a LVDS port (after all) it would be nice if it could be tested with the current default driver and if needed re-apply a patch for this LVDS port.
> 
> I hope I didn't miss any other purpose of this patch. Oh: I have programmed most cards to follow the existing pipe setup pre-programmed by the card BIOS since switching it almost never works on systems with the FDI interface (tested). Looks like programming can only be done reliably once the link training routines actually work (Which is not the case yet)..

Hi,

I was just trying to relax the constraints of hardcoded mappings between ports and pipes.

I don't remember the specifics for this one patch, but in general the idea was that I often saw machines where the pipe assignment from the BIOS didn't match our hardcoded constraints, and we went into attempting to reprogram everything to move a port from one pipe to another, which is not needed.

So I was just removing these constraints, and letting the driver read the existing mapping set in the BIOS from the hardware, and reuse it. It saves some time during the modesetting if we can do that. Currently the driver reprograms a lot of things on modeswitches and this results in a somewhat long delay with nothing on display. So if we can (safely) skip some of these steps, for example by keeping the pipe/ports assignments set in the BIOS, we should do it. And to achieve that more easily, we need to remove artificial constraints that are not related to any hardware limitation.

Now, you probably have a better understanding of the current state of the driver than I do, so if this does not make sense, you can abandon this patch.

Patch-set: 5
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>